fix: solving problems on accordion sidebar and restructuring page files
