+++ src/app/(private)/products/clients-pricing-table/page.tsx
@@ -32,7 +32,7 @@ import {
 
 import { useId, useState } from 'react';
 import { BsThreeDots } from 'react-icons/bs';
-import { ButtonFilter } from '../../../components/ButtonFilter';
+import { ButtonFilter } from '../../../../components/ButtonFilter';
 
 const range = (stop: number) =>
   Array.from({ length: stop })
