+++ src/app/(private)/products/standard-pricing-table/page.tsx
@@ -33,7 +33,7 @@ import {
 
 import { useId, useState } from 'react';
 import { BsThreeDots } from 'react-icons/bs';
-import { ButtonFilter } from '../../../components/ButtonFilter';
+import { ButtonFilter } from '../../../../components/ButtonFilter';
 
 const range = (stop: number) =>
   Array.from({ length: stop })
