+++ src/components/Sidebar.tsx
@@ -19,180 +19,10 @@ import { useWindowSize } from '@uidotdev/usehooks';
 import { motion } from 'framer-motion';
 import { usePathname } from 'next/navigation';
 import { useEffect, useState } from 'react';
-import { BiSolidFactory } from 'react-icons/bi';
-import { IoBagCheckSharp } from 'react-icons/io5';
 
-import {
-  MdAssignmentInd,
-  MdContacts,
-  MdDashboard,
-  MdDirectionsCar,
-  MdInsertChart,
-  MdMenu,
-  MdPaid,
-  MdPinDrop,
-  MdSettings,
-  MdStars,
-} from 'react-icons/md';
-import type { IconType } from 'react-icons';
+import { MdMenu } from 'react-icons/md';
 
-type Option = { icon: IconType; id: string; name: string } & (
-  | { path: string }
-  | { items: Array<{ id: string; name: string; path: string }> }
-);
-
-const options: Option[] = [
-  {
-    icon: MdDashboard,
-    id: Math.random().toString(36).substring(2),
-    name: 'Dashboard',
-    path: '/dashboard',
-  },
-  {
-    icon: MdInsertChart,
-    id: Math.random().toString(36).substring(2),
-    name: 'Painel Gerencial',
-    path: '/management-panel',
-  },
-  {
-    icon: MdContacts,
-    id: Math.random().toString(36).substring(2),
-    name: 'Clientes',
-    path: '/clients',
-  },
-  {
-    icon: MdPaid,
-    id: Math.random().toString(36).substring(2),
-    name: 'Notas Fiscais',
-    path: '/invoices',
-  },
-  {
-    icon: MdPinDrop,
-    id: Math.random().toString(36).substring(2),
-    items: [
-      {
-        id: Math.random().toString(36).substring(2),
-        name: 'Pedidos',
-        path: '/mobile/order-details',
-      },
-      {
-        id: Math.random().toString(36).substring(2),
-        name: 'Visitas',
-        path: '/mobile/visits',
-      },
-      {
-        id: Math.random().toString(36).substring(2),
-        name: 'Mapa de Visitas',
-        path: '/mobile/visits-map',
-      },
-      {
-        id: Math.random().toString(36).substring(2),
-        name: 'Estoque Cliente',
-        path: '/mobile/client-stock',
-      },
-      {
-        id: Math.random().toString(36).substring(2),
-        name: 'Estoque Distrubuidor',
-        path: '/mobile/distributor-stock',
-      },
-    ],
-    name: 'Mobile',
-  },
-  {
-    icon: IoBagCheckSharp,
-    id: Math.random().toString(36).substring(2),
-    items: [
-      {
-        id: Math.random().toString(36).substring(2),
-        name: 'Planejamento Clientes',
-        path: '/customer-planning',
-      },
-      {
-        id: Math.random().toString(36).substring(2),
-        name: 'Planejamento Roteiros',
-        path: '/visitas',
-      },
-    ],
-    name: 'Planejamentos',
-  },
-  {
-    icon: MdStars,
-    id: Math.random().toString(36).substring(2),
-    items: [
-      {
-        id: Math.random().toString(36).substring(2),
-        name: 'Lista de Produtos',
-        path: '/products',
-      },
-      {
-        id: Math.random().toString(36).substring(2),
-        name: 'Tabela de Preços Padrão',
-        path: '/standard-pricing-table',
-      },
-      {
-        id: Math.random().toString(36).substring(2),
-        name: 'Tabela de Preços Clientes',
-        path: '/clients-pricing-table',
-      },
-      {
-        id: Math.random().toString(36).substring(2),
-        name: 'Tabela de Preços Grupo',
-        path: '/group-pricing-table',
-      },
-      {
-        id: Math.random().toString(36).substring(2),
-        name: 'Solicitação Preço',
-        path: '/estoque-de-clientes',
-      },
-      {
-        id: Math.random().toString(36).substring(2),
-        name: 'Aprovações',
-        path: '/estoque-de-clientes',
-      },
-    ],
-    name: 'Produtos',
-  },
-  {
-    icon: BiSolidFactory,
-    id: Math.random().toString(36).substring(2),
-    items: [
-      {
-        id: Math.random().toString(36).substring(2),
-        name: 'Pedidos',
-        path: '/factory/orders',
-      },
-      {
-        id: Math.random().toString(36).substring(2),
-        name: 'Novos Pedidos',
-        path: '/factory/new-orders',
-      },
-      {
-        id: Math.random().toString(36).substring(2),
-        name: 'Estoque PPS',
-        path: '/factory/pps-stock',
-      },
-    ],
-    name: 'Fábrica',
-  },
-  {
-    icon: MdDirectionsCar,
-    id: Math.random().toString(36).substring(2),
-    name: 'Pedidos Parceiros',
-    path: '/partners-orders',
-  },
-  {
-    icon: MdAssignmentInd,
-    id: Math.random().toString(36).substring(2),
-    name: 'Usuários',
-    path: '/users',
-  },
-  {
-    icon: MdSettings,
-    id: Math.random().toString(36).substring(2),
-    name: 'Ferramentas',
-    path: '/tools',
-  },
-];
+import { options } from '../configs/sidebarOptions';
 
 export function Sidebar() {
   const pathname = usePathname();
@@ -301,7 +131,8 @@ export function Sidebar() {
           ) : (
             <Accordion
               key={option.id}
-              allowToggle
+              defaultIndex={[pathname.includes(option.rootPath) ? 0 : -1]}
+              allowMultiple
             >
               <AccordionItem
                 borderWidth={0}
@@ -355,7 +186,7 @@ export function Sidebar() {
                         gap={3}
                         h="2.5rem"
                         href={item.path}
-                        isActive={pathname.startsWith(item.path)}
+                        isActive={pathname.endsWith(item.path)}
                         justifyContent="flex-start"
                         px={9}
                         _active={{
@@ -372,7 +203,6 @@ export function Sidebar() {
                         <Box
                           as="span"
                           display={collapse ? 'none' : 'block'}
-                          transition="all 0.8s ease-in-out"
                         >
                           {item.name}
                         </Box>
